backend/requirements.txt
Flask==2.1.1
Flask-Cors==3.1.1
Flask-SocketIO==5.2.0
requests==2.26.0
python-dotenv==0.19.0
PyPDF2==1.26.0
python-docx==0.8.11